vk: Implement safe typeless transfer  - Used to transfer D32S8 data where it makes sense to use this variant  - On nvidia cards, it is very slow to move aspects from D24S8 probably due to the format being faked.    For this reason, the unsafe variant is used for both D16 and D24S8 to avoid the heavy performance loss

+++ rpcs3/Emu/RSX/VK/VKCompute.h
@@ -39,7 +39,7 @@ namespace vk
 			VkDescriptorSetLayoutCreateInfo infos = {};
 			infos.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
 			infos.pBindings = bindings.data();
-			infos.bindingCount = bindings.size();
+			infos.bindingCount = (u32)bindings.size();
 
 			CHECK_RESULT(vkCreateDescriptorSetLayout(*get_current_renderer(), &infos, nullptr, &m_descriptor_layout));
